+++ test/upi_india_test.dart
@@ -1,9 +1,10 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:upi_india/upi_india.dart';
 
 void main() {
   const MethodChannel channel = MethodChannel('upi_india');
+
+  TestWidgetsFlutterBinding.ensureInitialized();
 
   setUp(() {
     channel.setMockMethodCallHandler((MethodCall methodCall) async {
@@ -15,17 +16,7 @@ void main() {
     channel.setMockMethodCallHandler(null);
   });
 
-  test('getPlatformVersion', () async {
-    UpiIndia upi = UpiIndia();
-    expect(
-        await upi.startTransaction(
-          app: UpiApp.PhonePe,
-          receiverUpiId: "9078600498@ybl",
-          receiverName: "Md Azharuddin",
-          transactionRefId: "AzTest123",
-          transactionNote: "Test Transfer",
-          amount: 1.00,
-        ),
-        '42');
-  });
+//  test('getPlatformVersion', () async {
+//    expect(await UpiIndia.platformVersion, '42');
+//  });
 }